+++ src/Curves/NamedCurveFp.php
@@ -2,14 +2,23 @@
 
 namespace Mdanter\Ecc\Curves;
 
-use Mdanter\Ecc\CurveFp;
+use Mdanter\Ecc\Primitives\CurveFp;
 use Mdanter\Ecc\MathAdapterInterface;
 
 class NamedCurveFp extends CurveFp
 {
-
+    /**
+     * @var int|string
+     */
     private $name;
 
+    /**
+     * @param int|string $name
+     * @param int|string $prime
+     * @param int|string $a
+     * @param MathAdapterInterface $b
+     * @param MathAdapterInterface $adapter
+     */
     public function __construct($name, $prime, $a, $b, MathAdapterInterface $adapter)
     {
         $this->name = $name;
@@ -17,6 +26,9 @@ class NamedCurveFp extends CurveFp
         parent::__construct($prime, $a, $b, $adapter);
     }
 
+    /**
+     * @return int|string
+     */
     public function getName()
     {
         return $this->name;
