+++ src/EccFactory.php
@@ -5,6 +5,8 @@ namespace Mdanter\Ecc;
 use Mdanter\Ecc\Curves\NistCurve;
 use Mdanter\Ecc\Curves\SecgCurve;
 use Mdanter\Ecc\Math\MathAdapterFactory;
+use Mdanter\Ecc\Math\NumberTheory;
+use Mdanter\Ecc\Primitives\CurveFp;
 
 /**
  * Static factory class providing factory methods to work with NIST and SECG recommended curves.
@@ -31,7 +33,7 @@ class EccFactory
      * Contains useful modular/polynomial functions
      *
      * @param  MathAdapterInterface      $adapter [optional] Defaults to the return value EccFactory::getAdapter().
-     * @return \Mdanter\Ecc\NumberTheory
+     * @return \Mdanter\Ecc\Math\NumberTheory
      */
     public static function getNumberTheory(MathAdapterInterface $adapter = null)
     {
@@ -67,7 +69,7 @@ class EccFactory
      * @param  int|string                    $a
      * @param  int|string                    $b
      * @param  MathAdapterInterface          $adapter [optional] Defaults to the return value of EccFactory::getAdapter().
-     * @return \Mdanter\Ecc\CurveFpInterface
+     * @return \Mdanter\Ecc\Primitives\CurveFpInterface
      */
     public static function createCurve($prime, $a, $b, MathAdapterInterface $adapter = null)
     {
