+++ src/Math/MathAdapterFactory.php
@@ -8,11 +8,18 @@ class MathAdapterFactory
 {
     private static $forcedAdapter = null;
 
+    /**
+     * @param MathAdapterInterface $adapter
+     */
     public static function forceAdapter(MathAdapterInterface $adapter = null)
     {
         self::$forcedAdapter = $adapter;
     }
 
+    /**
+     * @param bool $debug
+     * @return DebugDecorator|MathAdapterInterface|null
+     */
     public static function getAdapter($debug = false)
     {
         if (self::$forcedAdapter !== null) {
@@ -27,6 +34,10 @@ class MathAdapterFactory
         return self::wrapAdapter($adapter, (bool) $debug);
     }
 
+    /**
+     * @param bool $debug
+     * @return DebugDecorator|MathAdapterInterface
+     */
     public static function getGmpAdapter($debug = false)
     {
         if (self::canLoad('gmp')) {
@@ -36,7 +47,10 @@ class MathAdapterFactory
         throw new \RuntimeException('Please install GMP extension.');
     }
 
-    private static function getAdapterClass($extension = null)
+    /**
+     * @return string
+     */
+    private static function getAdapterClass()
     {
         if (self::canLoad('gmp')) {
             return '\Mdanter\Ecc\Math\Gmp';
@@ -45,11 +59,20 @@ class MathAdapterFactory
         throw new \RuntimeException('Please install GMP extension.');
     }
 
+    /**
+     * @param $target
+     * @return bool
+     */
     private static function canLoad($target)
     {
         return extension_loaded($target);
     }
 
+    /**
+     * @param MathAdapterInterface $adapter
+     * @param $debug
+     * @return DebugDecorator|MathAdapterInterface
+     */
     private static function wrapAdapter(MathAdapterInterface $adapter, $debug)
     {
         if ($debug === true) {
