+++ src/Primitives/PointInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mdanter\Ecc;
+namespace Mdanter\Ecc\Primitives;
 
 /**
  * *********************************************************************
@@ -25,6 +25,8 @@ namespace Mdanter\Ecc;
  * OTHER DEALINGS IN THE SOFTWARE.
  * ***********************************************************************
  */
+use Mdanter\Ecc\Primitives\CurveFpInterface;
+use Mdanter\Ecc\Infinity;
 
 /**
  * This is the contract for implementing Point, which encapsulates entities
