+++ src/Random/HmacRandomNumberGenerator.php
@@ -3,7 +3,7 @@
 namespace Mdanter\Ecc\Random;
 
 use Mdanter\Ecc\MathAdapterInterface;
-use Mdanter\Ecc\PrivateKeyInterface;
+use Mdanter\Ecc\Crypto\PrivateKeyInterface;
 use Mdanter\Ecc\RandomNumberGeneratorInterface;
 use Mdanter\Ecc\Util\NumberSize;
