+++ src/Serializer/PrivateKey/PrivateKeySerializerInterface.php
@@ -2,7 +2,7 @@
 
 namespace Mdanter\Ecc\Serializer\PrivateKey;
 
-use Mdanter\Ecc\PrivateKeyInterface;
+use Mdanter\Ecc\Crypto\PrivateKeyInterface;
 
 interface PrivateKeySerializerInterface
 {
