+++ src/Serializer/PublicKey/Der/Formatter.php
@@ -5,8 +5,8 @@ namespace Mdanter\Ecc\Serializer\PublicKey\Der;
 use FG\ASN1\Universal\Sequence;
 use FG\ASN1\Universal\ObjectIdentifier;
 use FG\ASN1\Universal\BitString;
-use Mdanter\Ecc\PointInterface;
-use Mdanter\Ecc\PublicKeyInterface;
+use Mdanter\Ecc\Primitives\PointInterface;
+use Mdanter\Ecc\Crypto\PublicKeyInterface;
 use Mdanter\Ecc\MathAdapterInterface;
 use Mdanter\Ecc\Curves\NamedCurveFp;
 use Mdanter\Ecc\Serializer\Util\CurveOidMapper;
